Set up unit tests for `lpm` and the git module
*  Decided to test the git module by creating an empty (init --bare) repository and turning it into a tarball.  That way, it can be used as a reference in unit tests to assert that an
operation who's sole purpose was to create a files suceeded.
*  Added a `fixtures` folder containing the tarball archive of the bare repo (`git_bare.tar`)
*  Created an extension of `unittest.TestCase` that stubs out what we would need to add to parent class to support the use of those fixtures.
*  Stubbed two new assertions (`assertDirectoryEquals` and `assertDirectoryContains`) that should be implmented to perform the comparison.

The idea is that `tests.test_git.TestGitModule` would inherit from `lpm_unittest.LPMTestCase` instead of `unittest.TestCase`.  That way, it can contain a test that executes a
function where the expected output isn't verified by a return value, but by a comparison of a directory tree.  An assertion could be made with `os.walk`, but the test would still
have to hard-code the expected result and that doesn't do anything for legibility and becomes a bit of a maintenance headache.  Better would be if we could describe the expected
result in a file that could be included with the unittest and compare the actual directory structure (via `os.walk`) to that.  This allows us to just update the expected result by
generating it once, confirming it by hand and just commmitting it.

I should point that this has been inspired by the need to ensure that my extension to the git python module is creating the correct kinds of bare repsoitories and things.