Add `assertDirectoryEquals
*  Added a functional `lpm_unittest.LPMTestCase.assertDirectoryEqual` method.
*  Finished the `fixtures` capabiltiy of `LPMTestsCase` so that files can be loaded into the TestCase from
   a `/fixtures/` directory.
*  `quarks.map_path` now supports a `filters` argument

## assertDirectoryEqual

This application is going to be doing a lot of work editing and altering the
filesystem where it is running.  As such, our unit tests have been given the
ability to assert that a directory structure from `path` matches one that has
been specified as a `tarfile.TarFile` object in the module's fixtures configuration.